Eliminate dnssec-enable option for bind version 9.16 and newer.
In bind version 9.16.0 the dnssec-enable option was made obsolete
and in 9.18.0 the option was entirely removed. Patch check
installed bind version and set variable bind_dnssec_enable
to false if version 9.16 and newer. Since the name of the package
is different in operating systems families i added variable bind_package.